+++ app/assets/stylesheets/application.scss
@@ -35,6 +35,7 @@ $theme-colors: $colors;
 
 @import "bootstrap/scss/bootstrap";
 @import "./scores_widget";
+@import "./autocomplete";
 @import "./devise";
 @import url("https://fonts.googleapis.com/css?family=Crimson+Text:400&subset=latin");
 @import url("https://fonts.googleapis.com/css?family=Pathway+Gothic+One&subset=latin");
