+++ app/javascript/packs/application.js
@@ -19,8 +19,10 @@ import "stylesheets/application.scss"
 
 import { Application } from "stimulus"
 import { definitionsFromContext } from "stimulus/webpack-helpers"
+
 import Turbolinks from "turbolinks"
 import Rails from "jquery-ujs"
+import {} from "jquery-ui/ui/widgets/autocomplete"
 
 Turbolinks.start()
