+++ package.json
@@ -6,6 +6,7 @@
     "@stimulus/webpack-helpers": "^1.1.1",
     "bootstrap": "4.3.1",
     "jquery": "^3.4.1",
+    "jquery-ui": "^1.12.1",
     "jquery-ujs": "^1.2.2",
     "popper.js": "^1.15.0",
     "stimulus": "^1.1.1",
